requirements.txt
PyQt5>=5.15
pyFluidSynth>=1.3
python-rtmidi>=1.5
